requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
decorator==4.3.0
Django==2.1.4
djangorestframework==3.9.0
future==0.17.1
geocoder==1.38.1
idna==2.8
opencage==1.1.6
pkg-resources==0.0.0
psycopg2==2.7.6.1
pytz==2018.7
ratelim==0.1.6
requests==2.21.0
six==1.12.0
urllib3==1.24.1
